Modifies at weekend.
Foreword content comes from Foreword.txt
Added the none.bmp to display when the result is null.
Changed the Teach, Link and Foreword to Inherit from the Base.
Added the control buttons to Base.
Moved uiCHSTRAS_091 to uiCHSTRAS_091_Base
Modified the Test file to fit the new constructor.
Changed the File in BT to make it to be more expansibility.
Added the new method to handle txt file and the string from database.
Added the query method to simplify the database query.
Modified the moveToNextString to distinct from the null result.
Added the getPageId methord to get page id from Pratice Unit ID.